fix(lora): run adapter matmuls in the activation dtype
The LoRA/LoKr factors are kept in float32 as master weights (fp16 params
underflow the gradient at typical learning rates), but they were also being
multiplied in float32. fp16 @ fp32 promotes, and since every patched layer
feeds the next block, one un-cast factor drags the whole transformer to
float32 and silently cancels --dtype: ~20% slower per step on fp16, with no
error and no visible sign in the config.

Cast the factors to x.dtype in LoRALinear, FusedLoRALinear and LoKrLinear
before the matmuls. The factors themselves stay float32, so training
precision is unchanged; casting an (in_dims, r) tensor is free next to the
base GEMM.

Adds tests/lora/test_lora_layer_dtype.py as a regression guard across
float16/bfloat16/float32 for all three layer types.

+++ src/mflux/models/common/lora/layer/linear_lora_layer.py
@@ -49,5 +49,10 @@ class LoRALinear(nn.Module):
 
     def __call__(self, x):
         base_out = self.linear(x)
-        lora_out = mx.matmul(mx.matmul(x, self.lora_A), self.lora_B)
+        # The factors stay float32 as master weights (fp16 LoRA params underflow the
+        # gradient at typical learning rates), but the matmuls must run in the activation
+        # dtype: fp16 @ fp32 promotes to fp32, and since every patched layer feeds the
+        # next block, one un-cast factor drags the whole transformer to fp32 and silently
+        # cancels --dtype. Casting (in_dims, r) is free next to the base GEMM.
+        lora_out = mx.matmul(mx.matmul(x, self.lora_A.astype(x.dtype)), self.lora_B.astype(x.dtype))
         return base_out + self.scale * lora_out
